+++ internal/stats/series.go
@@ -18,9 +18,8 @@ func (b *bucket) add(x bucket) {
 
 // Series contains logic of accumulating time series data.
 type Series struct {
-	total   bucket
-	current bucket
 	buckets []bucket
+	zero    []bucket
 
 	duration time.Duration
 	touched  time.Time
@@ -40,14 +39,14 @@ func NewSeries(d time.Duration, n int) *Series {
 	}
 	return &Series{
 		buckets:  make([]bucket, n),
+		zero:     make([]bucket, n),
 		duration: d,
 	}
 }
 
 // Add adds given value x at the time moment of now.
 func (s *Series) Add(now time.Time, x float64) {
-	s.rotate(now)
-	s.current.add(bucket{
+	s.buckets[s.rotate(now)].add(bucket{
 		sum: x,
 		cnt: 1,
 	})
@@ -55,8 +54,13 @@ func (s *Series) Add(now time.Time, x float64) {
 
 // Get returns accumulated data available at the moment of now.
 func (s *Series) Get(now time.Time) (sum float64, cnt int64) {
-	s.rotate(now)
-	return s.total.sum, s.total.cnt
+	var (
+		total = bucket{}
+	)
+	for i := s.rotate(now); i < len(s.buckets); i++ {
+		total.add(s.buckets[i])
+	}
+	return total.sum, total.cnt
 }
 
 // SumPer returns rate of accumulated data available at the moment of now.
@@ -65,10 +69,10 @@ func (s *Series) SumPer(now time.Time, period time.Duration) float64 {
 	return sum * float64(period) / float64(s.duration)
 }
 
-func (s *Series) rotate(now time.Time) {
+func (s *Series) rotate(now time.Time) (nowBucket int) {
 	if s.touched.IsZero() {
 		s.touched = now
-		return
+		return 0
 	}
 
 	var (
@@ -77,49 +81,48 @@ func (s *Series) rotate(now time.Time) {
 		span  = s.duration / time.Duration(n)
 		shift = int(d / span)
 	)
+
 	if shift == 0 {
-		return
+		return 0
 	}
 
 	// Slide window strictly on the span size grid.
-	s.touched = s.touched.Add(time.Duration(shift) * span)
-
-	if shift > n {
-		s.reset()
-		return
+	// only to the future
+	if shift > 0 {
+		s.touched = s.touched.Add(time.Duration(shift) * span)
 	}
 
-	current := s.current
-	s.current = bucket{}
-	s.total = current
-
-	for i := n - 1; i > 0; i-- {
-		var prev bucket
-		if i-shift >= 0 {
-			prev = s.buckets[i-shift]
-		}
-		s.total.add(prev)
-		s.buckets[i] = prev
+	if shift >= n || shift < -n+1 {
+		copy(s.buckets, s.zero)
+		return 0
 	}
-	s.buckets[shift-1] = current
+
+	// bucket in the past
+	if shift < 0 {
+		return -shift
+	}
+
+	// shift buckets front to tail
+	copy(s.buckets[shift:], s.buckets)
+
+	// zerois skips
+	copy(s.buckets, s.zero[n-shift:])
+
+	return shift - 1
 }
 
 func (s *Series) String() string {
-	var buf bytes.Buffer
-	fmt.Fprintf(&buf, "%0.f %d ", s.current.sum, s.current.cnt)
+	var (
+		buf   bytes.Buffer
+		total = bucket{}
+	)
 	for i, b := range s.buckets {
 		if i > 0 {
 			fmt.Fprint(&buf, ", ")
 		}
 		fmt.Fprintf(&buf, "[%0.f %d]", b.sum, b.cnt)
+		total.add(b)
 	}
-	fmt.Fprintf(&buf, " = %0.f %d\n", s.total.sum, s.total.cnt)
+	fmt.Fprintf(&buf, " = %0.f %d\n", total.sum, total.cnt)
 	return buf.String()
-}
-
-func (s *Series) reset() {
-	s.current = bucket{}
-	for i := range s.buckets {
-		s.buckets[i] = bucket{}
-	}
 }